+++ aries-site/src/examples/components/tag/TagAnatomy.js
@@ -1,8 +1,8 @@
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
-import { Box, Grid, Diagram, Stack, Text, ThemeContext } from 'grommet';
+import { Box, Button, Grid, Diagram, Stack, Text, ThemeContext } from 'grommet';
 import { FormClose } from 'grommet-icons';
-import { Annotation } from '../namevaluelist/NameValueListAnatomy';
+import { Annotation } from '../../../layouts';
 
 const color = 'border';
 const anchor = 'vertical';
@@ -55,11 +55,14 @@ export const AnatomyGrid = ({ ...rest }) => (
 
 export const AnatomyBox = ({ id, nameId, valueId, closeId }) => {
   const theme = useContext(ThemeContext);
+  const boxProps = {
+    direction: 'row',
+    gap: 'xxsmall',
+  };
 
   return (
     <Box
-      direction="row"
-      gap="xxsmall"
+      {...boxProps}
       id={id}
       border={theme.tag.border}
       pad={theme.tag.pad}
@@ -67,15 +70,20 @@ export const AnatomyBox = ({ id, nameId, valueId, closeId }) => {
     >
       <Box id={nameId}>Location</Box>
       {valueId && closeId && (
-        <>
-          <Box>:</Box>
+        <Box {...boxProps}>
+          <Box>: </Box>
           <Box id={valueId}>
             <Text {...theme.tag.value}>NY_USA</Text>
           </Box>
-          <Box id={closeId}>
-            <FormClose />
-          </Box>
-        </>
+          <Button
+            id={closeId}
+            icon={<FormClose />}
+            hoverIndicator
+            focusIndicator
+            plain
+            style={{ borderRadius: '50%' }}
+          />
+        </Box>
       )}
     </Box>
   );
@@ -85,25 +93,21 @@ export const TagAnatomy = () => {
   const theme = useContext(ThemeContext);
 
   return (
-    <Stack>
+    <Stack interactiveChild="first">
       <AnatomyGrid>
         {/* Empty Box occupies first cell of grid. Alternatively 
         could use gridAreas, but felt heavy for this need. */}
         <Box />
         <Annotation id="border-annotation" target="1" />
         <Box />
-        <Box alignSelf="center">
-          <Annotation id="name-annotation" target="2" />
-        </Box>
+        <Annotation alignSelf="center" id="name-annotation" target="2" />
         <AnatomyBox
           id="border"
           nameId="tagName-2"
           valueId="value"
           closeId="close"
         />
-        <Box alignSelf="center">
-          <Annotation id="close-annotation" target="4" />
-        </Box>
+        <Annotation alignSelf="center" id="close-annotation" target="4" />
         <Box />
         {/* non t-shirt size necessary for diagram usecase for annotation
           to vertically align with item */}
